Re-jigger the project because Eclipse is needy. This also created a main activity and retargeted the project for Jelly Bean.
Re-jigger the project because Eclipse is needy. This also created a main activity and retargeted the project for Jelly Bean.

You should, now, be able to import the project into Eclipse by using File > Import… > General > Existing Projects into Workspace and selecting the CHECKED OUT ROOT of the repo as the root directory (i.e., ProjectTeamRocket, not FindMyThings). FindMyThings will appear in the Projects box and hit next. :)